chore: Refactor Boxes, context menu
