+++ src/javascript/JContent/EditFrame/Boxes/index.js
@@ -1,0 +1,1 @@
+export {Boxes} from './Boxes';
